refactor(db): extract search palette to db/src/palette.rs
