+++ db/src/lib.rs
@@ -17,6 +17,7 @@ pub mod helpers;
 pub mod indexer;
 pub mod library_layout;
 pub mod metadata_overrides;
+pub mod palette;
 pub mod pool;
 pub mod queries;
 pub mod scanner;
@@ -41,6 +42,7 @@ pub use covers::{covers_dir, get_cover, get_last_modified_epoch};
 pub use discovery::*;
 pub use helpers::{build_fts_match, sanitize_fts_query};
 pub use metadata_overrides::*;
+pub use palette::*;
 pub use pool::*;
 pub use queries::*;
 pub use settings::*;
